refactor plugin resource call with/without ds
