+++ pkg/plugins/plugincontext/plugincontext.go
@@ -43,11 +43,34 @@ type Provider struct {
 // Get allows getting plugin context by its ID. If datasourceUID is not empty string
 // then PluginContext.DataSourceInstanceSettings will be resolved and appended to
 // returned context.
-func (p *Provider) Get(ctx context.Context, pluginID string, datasourceUID string, user *models.SignedInUser, skipCache bool) (backend.PluginContext, bool, error) {
-	pc := backend.PluginContext{}
+func (p *Provider) Get(ctx context.Context, pluginID string, user *models.SignedInUser) (backend.PluginContext, bool, error) {
+	return p.pluginContext(ctx, pluginID, user)
+}
+
+// GetWithDataSource allows getting plugin context by its ID and PluginContext.DataSourceInstanceSettings will be
+// resolved and appended to the returned context.
+func (p *Provider) GetWithDataSource(ctx context.Context, pluginID string, user *models.SignedInUser, ds *models.DataSource) (backend.PluginContext, bool, error) {
+	pCtx, exists, err := p.pluginContext(ctx, pluginID, user)
+	if err != nil {
+		return pCtx, exists, err
+	}
+
+	datasourceSettings, err := adapters.ModelToInstanceSettings(ds, p.decryptSecureJsonDataFn(ctx))
+	if err != nil {
+		return pCtx, exists, errutil.Wrap("Failed to convert datasource", err)
+	}
+	pCtx.DataSourceInstanceSettings = datasourceSettings
+
+	return pCtx, true, nil
+}
+
+const pluginSettingsCacheTTL = 5 * time.Second
+const pluginSettingsCachePrefix = "plugin-setting-"
+
+func (p *Provider) pluginContext(ctx context.Context, pluginID string, user *models.SignedInUser) (backend.PluginContext, bool, error) {
 	plugin, exists := p.pluginStore.Plugin(ctx, pluginID)
 	if !exists {
-		return pc, false, nil
+		return backend.PluginContext{}, false, nil
 	}
 
 	jsonData := json.RawMessage{}
@@ -59,18 +82,18 @@ func (p *Provider) Get(ctx context.Context, pluginID string, datasourceUID strin
 		// models.ErrPluginSettingNotFound is expected if there's no row found for plugin setting in database (if non-app plugin).
 		// If it's not this expected error something is wrong with cache or database and we return the error to the client.
 		if !errors.Is(err, models.ErrPluginSettingNotFound) {
-			return pc, false, errutil.Wrap("Failed to get plugin settings", err)
+			return backend.PluginContext{}, false, errutil.Wrap("Failed to get plugin settings", err)
 		}
 	} else {
 		jsonData, err = json.Marshal(ps.JSONData)
 		if err != nil {
-			return pc, false, errutil.Wrap("Failed to unmarshal plugin json data", err)
+			return backend.PluginContext{}, false, errutil.Wrap("Failed to unmarshal plugin json data", err)
 		}
 		decryptedSecureJSONData = p.pluginSettingsService.DecryptedValues(ps)
 		updated = ps.Updated
 	}
 
-	pCtx := backend.PluginContext{
+	return backend.PluginContext{
 		OrgID:    user.OrgId,
 		PluginID: plugin.ID,
 		User:     adapters.BackendUserFromSignedInUser(user),
@@ -79,25 +102,8 @@ func (p *Provider) Get(ctx context.Context, pluginID string, datasourceUID strin
 			DecryptedSecureJSONData: decryptedSecureJSONData,
 			Updated:                 updated,
 		},
-	}
-
-	if datasourceUID != "" {
-		ds, err := p.dataSourceCache.GetDatasourceByUID(ctx, datasourceUID, user, skipCache)
-		if err != nil {
-			return pc, false, errutil.Wrap("Failed to get datasource", err)
-		}
-		datasourceSettings, err := adapters.ModelToInstanceSettings(ds, p.decryptSecureJsonDataFn(ctx))
-		if err != nil {
-			return pc, false, errutil.Wrap("Failed to convert datasource", err)
-		}
-		pCtx.DataSourceInstanceSettings = datasourceSettings
-	}
-
-	return pCtx, true, nil
+	}, true, nil
 }
-
-const pluginSettingsCacheTTL = 5 * time.Second
-const pluginSettingsCachePrefix = "plugin-setting-"
 
 func (p *Provider) getCachedPluginSettings(ctx context.Context, pluginID string, user *models.SignedInUser) (*pluginsettings.DTO, error) {
 	cacheKey := pluginSettingsCachePrefix + pluginID
