+++ pkg/services/live/liveplugin/plugin.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/grafana/grafana/pkg/services/datasources"
+	"github.com/grafana/grafana/pkg/util/errutil"
+
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/plugins/plugincontext"
 	"github.com/grafana/grafana/pkg/services/live/orgchannel"
@@ -60,15 +63,26 @@ func (p *NumLocalSubscribersGetter) GetNumLocalSubscribers(channelID string) (in
 }
 
 type ContextGetter struct {
-	PluginContextProvider *plugincontext.Provider
+	pluginContextProvider *plugincontext.Provider
+	dataSourceCache       datasources.CacheService
 }
 
-func NewContextGetter(pluginContextProvider *plugincontext.Provider) *ContextGetter {
+func NewContextGetter(pluginContextProvider *plugincontext.Provider, dataSourceCache datasources.CacheService) *ContextGetter {
 	return &ContextGetter{
-		PluginContextProvider: pluginContextProvider,
+		pluginContextProvider: pluginContextProvider,
+		dataSourceCache:       dataSourceCache,
 	}
 }
 
 func (g *ContextGetter) GetPluginContext(ctx context.Context, user *models.SignedInUser, pluginID string, datasourceUID string, skipCache bool) (backend.PluginContext, bool, error) {
-	return g.PluginContextProvider.Get(ctx, pluginID, datasourceUID, user, skipCache)
+	if datasourceUID == "" {
+		return g.pluginContextProvider.Get(ctx, pluginID, user)
+	}
+
+	ds, err := g.dataSourceCache.GetDatasourceByUID(ctx, datasourceUID, user, skipCache)
+	if err != nil {
+		return backend.PluginContext{}, false, errutil.Wrap("Failed to get datasource", err)
+	}
+	return g.pluginContextProvider.GetWithDataSource(ctx, pluginID, user, ds)
+
 }
